Sophgo/DwMac4SnpDxe: Fix the data length of `RecycledTxBuf`
- The `RecycledTxBuf` should be UINT64*

- Enable build ethernet driver by default

+++ Silicon/Sophgo/Drivers/Net/DwMac4SnpDxe/DwMac4SnpDxe.c
@@ -1073,7 +1073,7 @@ SnpTransmit (
   DMA_DESCRIPTOR                    *TxDescriptor;
   DMA_DESCRIPTOR                    *TxDescriptorMap;
   UINT8                             *EthernetPacket;
-  UINT8                             *Tmp;
+  UINT64                            *Tmp;
   EFI_STATUS                        Status;
   EFI_PHYSICAL_ADDRESS              TxBufferPhysAddress;
   UINT32                            Index;
